+++ frontend/package.json
@@ -23,7 +23,8 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-infinite-scroller": "^1.2.6",
-    "react-router-dom": "^6.22.3"
+    "react-router-dom": "^6.22.3",
+    "swr": "^2.2.5"
   },
   "devDependencies": {
     "@types/node": "^20.11.28",
